cnf split into a separate file
